requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-google==10.3.0
soda-core-bigquery==3.0.45
soda-core-scientific==3.0.45
astronomer-cosmos[dbt-bigquery]
protobuf